requirements.txt
absl-py==0.8.0
astor==0.8.0
attrs==19.3.0
certifi==2019.9.11
click==7.1.1
click-plugins==1.1.1
cligj==0.5.0
cycler==0.10.0
decorator==4.4.1
efficientnet==1.0.0
Fiona==1.8.13.post1
gast==0.3.2
google-pasta==0.1.7
grpcio==1.16.1
h5py==2.9.0
image-classifiers==1.0.0
imageio==2.6.1
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
munch==2.5.0
networkx==2.4
numpy==1.17.2
opencv-python==4.2.0.34
pandas==0.25.1
Pillow==7.0.0
protobuf==3.9.2
pyparsing==2.4.6
pyproj==2.6.0
pyshp==2.1.0
python-dateutil==2.8.0
python-magic==0.4.15
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.1
segmentation-models==1.0.1
Shapely==1.7.0
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
Werkzeug==0.16.0
wrapt==1.11.2
